+++ com.oracle.truffle.r.test/src/com/oracle/truffle/r/test/parser/TestParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2015, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -187,7 +187,7 @@ public class TestParser extends TestBase {
                 RParserFactory.Parser<RSyntaxNode> parser = RParserFactory.getParser();
                 try {
                     source = RSource.fromFile(file);
-                    parser.script(source, new RASTBuilder(), null);
+                    parser.script(source, new RASTBuilder(true), null);
                 } catch (Throwable e) {
                     errorCount++;
                     Throwable t = e;